+++ packages/core/src/behaviors/export.ts
@@ -1,0 +1,8 @@
+/**
+ * @description Exports core behaviors modules.
+ * @file src/behaviors/export.ts
+ * @author Luca Liguori
+ * @created 2026-03-04
+ * @version 1.0.0
+ * @license Apache-2.0
+ */