+++ packages/core/src/helpers.ts
@@ -149,10 +149,9 @@ export async function addVirtualDevice(
 export async function addVirtualDevices(matterbridge: Matterbridge, aggregatorEndpoint: Endpoint<AggregatorEndpoint>): Promise<void> {
   // istanbul ignore next - No test for now cause is just a way to easily add new devices for testing purposes without using dynamic plugin
   if (hasParameter('experimental') && matterbridge.bridgeMode === 'bridge' && aggregatorEndpoint) {
-    // 36353036363234353431
-    const lock = new MatterbridgeEndpoint(doorLockDevice, { id: 'system_lock' });
-    lock.createDefaultBridgedDeviceBasicInformationClusterServer(
-      'Matterbridge System Lock',
+    const lockPin = new MatterbridgeEndpoint(doorLockDevice, { id: 'door_lock_pin' });
+    lockPin.createDefaultBridgedDeviceBasicInformationClusterServer(
+      'Matterbridge Pin Lock',
       'sn_system_lock',
       0xfff1,
       'Matterbridge',
@@ -160,9 +159,23 @@ export async function addVirtualDevices(matterbridge: Matterbridge, aggregatorEn
       20000,
       '2.0.0',
     );
-    lock.createPinDoorLockClusterServer();
-    lock.addRequiredClusterServers();
-    await aggregatorEndpoint.add(lock);
+    lockPin.createPinDoorLockClusterServer();
+    lockPin.addRequiredClusterServers();
+    await aggregatorEndpoint.add(lockPin);
+
+    const lockUserPin = new MatterbridgeEndpoint(doorLockDevice, { id: 'door_lock_user_pin' });
+    lockUserPin.createDefaultBridgedDeviceBasicInformationClusterServer(
+      'Matterbridge User Pin Lock',
+      'sn_system_lock',
+      0xfff1,
+      'Matterbridge',
+      'Matterbridge Virtual Device',
+      20000,
+      '2.0.0',
+    );
+    lockUserPin.createUserPinDoorLockClusterServer();
+    lockUserPin.addRequiredClusterServers();
+    await aggregatorEndpoint.add(lockUserPin);
   }
   if (matterbridge.virtualMode !== 'disabled' && matterbridge.bridgeMode === 'bridge' && aggregatorEndpoint) {
     matterbridge.log.notice(`Creating virtual devices for Matterbridge server node...`);